Add global styles to new styled elements, moved all components to use custom elements rather than styled.xx
